Refactor pkg_expose to leverage pkg_exports
This changes the accepted values for pkg_expose from a numeric literal
to a string representing a key in pkg_exports containing a valid port
number. The contents of EXPOSES will still contain the numeric
literal so this is not a breaking change for the Habitat runtime,
just a breaking change for hab-plan-build users.

* A numeric literal is no longer a valid pkg_expose
* The default value of the pkg_export must be a valid port number
  or package generation will halt
* Package generation will fail if the pkg_export is missing

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>